anecdotes-redux

+++ part4/part4_01/pathcontroller/blogroutes.js
@@ -5,7 +5,7 @@ const Blog= require('../models/blogs')
 const User= require('../models/user')
 const mongoose = require('mongoose')
 const jwt = require('jsonwebtoken')
-mongoose.set("useFindAndModify", true)
+mongoose.set("useFindAndModify", false)
 blogRouter.use(middleware.tokenExtractor)
 
 
@@ -16,16 +16,23 @@ blogRouter.get('/', async (request, response) => {
 })
 
 blogRouter.post('/', async (request, response) => {
+    console.log("here")
     const body= request.body
+    console.log(body)
+    console.log("Potatogear")
     const decodedToken = jwt.verify(request.token,process.env.SECRET)
+
     if (!request.token || !decodedToken.id){
         return response.status(401).json({error: 'token missing or invalid'})
     }
     if (!body.likes){
-        blog.likes=0
+        body.likes=0
     }
+    console.log(body)
+
     if(body.title && body.author && body.url){
         const user = await User.findOne({name:body.author.toLowerCase()})
+        console.log(user)
         const blog = new Blog({
             title: body.title,
             author: body.author.toLowerCase(),
@@ -47,7 +54,26 @@ blogRouter.post('/', async (request, response) => {
 
 blogRouter.delete('/:id', async (request, response) => {
     const id_ = request.params.id
-    await Blog.findByIdAndRemove(id_)
+    const decodedToken = jwt.verify(request.token,process.env.SECRET)
+    if (!request.token || !decodedToken.id){
+        return response.status(401).json({error: 'token missing or invalid'})
+    }
+    const blog = await Blog.findById(id_)
+    if(blog) {
+
+        const ver = await jwt.verify(request.token, process.env.SECRET)
+
+
+        if (ver.id === blog.user.toString()) {
+            await Blog.findByIdAndRemove(id_)
+        } else {
+            return response.status(401).json({error: 'You dont have permission to delete this blog'})
+        }
+    }else{
+        return response.status(401).json({error: 'No such blog exists'})
+    }
+
+
     response.status(204).end()
 })
